TIP-870: Move pdf specs to Akeneo\Enrichement
