+++ phpspec.yml.dist
@@ -242,11 +242,6 @@ suites:
         psr4_prefix: Pim\Bundle\NotificationBundle
         spec_path: src/Pim/Bundle/NotificationBundle
         src_path: src/Pim/Bundle/NotificationBundle
-    PdfGeneratorBundle:
-        namespace: Pim\Bundle\PdfGeneratorBundle
-        psr4_prefix: Pim\Bundle\PdfGeneratorBundle
-        spec_path: src/Pim/Bundle/PdfGeneratorBundle
-        src_path: src/Pim/Bundle/PdfGeneratorBundle
     ReferenceDataBundle:
         namespace: Pim\Bundle\ReferenceDataBundle
         psr4_prefix: Pim\Bundle\ReferenceDataBundle
