+++ src/Akeneo/Pim/Enrichment/Bundle/spec/Controller/Ui/ProductControllerSpec.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spec\Pim\Bundle\PdfGeneratorBundle\Controller;
+namespace spec\Akeneo\Pim\Enrichment\Bundle\Controller\Ui;
 
 use PhpSpec\ObjectBehavior;
 use Akeneo\Pim\Enrichment\Bundle\PdfGeneration\Renderer\RendererRegistry;
